% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_candy.R
\name{plot_candy}
\alias{plot_candy}
\title{Plot Candy Popularity}
\usage{
plot_candy(df, n = 10, candy_name_col, rank_col)
}
\arguments{
\item{df}{a dataframe containing information on the candies}

\item{n}{the number of top candies to plot.  Default is set at 10}

\item{candy_name_col}{the name of the column in \code{df} that specifies the names of the candies}

\item{rank_col}{the name of the column in \code{df} that specifies how to rank the candies}
}
\value{
a ggplot object
}
\description{
This function produces a bar plot of the most popular Halloween candies
}
\examples{
plot_candy(candy_data, candy_name_col = "competitorname", rank_col = "winpercent")
}

